Separate BIOS stub from PIC model
Issue #32
