+++ int-advtech.h
@@ -1,3 +1,7 @@
+//------------------------------------------------------------------------------
+// EMU86 - Advantech interrupt controller
+//------------------------------------------------------------------------------
+
 #pragma once
 
 #include "op-common.h"
