+++ int-elks.h
@@ -1,3 +1,7 @@
+//------------------------------------------------------------------------------
+// EMU86 - ELKS interrupt controller (8259)
+//------------------------------------------------------------------------------
+
 #pragma once
 
 // Interrupt lines
